requirements.txt
pyCryptodome==3.23.0